delete company front UI, password change layout, login password preview

+++ XYZEngineeringProject.UI/src/app/components/settings/settings.component.ts
@@ -1,3 +1,5 @@
+import { User } from './../../models/user.model';
+import { RolesDB } from './../../models/roles.enum';
 import { CompanyService } from './../../services/company/company.service';
 import { AuthorizationService } from 'src/app/services/authorization/authorization.service';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -13,7 +15,12 @@ import { TranslateService } from '@ngx-translate/core';
 export class SettingsComponent implements OnInit {
   
   currentUserRole: any;
-  currentUserId: string;
+  currentUser: User;
+  canModifyRoles: RolesDB[] = [
+    RolesDB.Admin,
+    RolesDB.Moderator,
+    RolesDB.Management
+  ];
 
   constructor(
     private router: Router,
@@ -26,9 +33,7 @@ export class SettingsComponent implements OnInit {
   ngOnInit(): void {
     this.authorizationService.currentUser().subscribe({
       next: (res) => {
-        this.currentUserId = res.id;
-        //TODO przypisać role
-
+        this.currentUser = res;
       }
     })
   }
@@ -39,16 +44,47 @@ export class SettingsComponent implements OnInit {
   }
 
   deleteCompany(){
-    if (confirm(this.translate.instant('Alert.deleteCompany'))) {
-      this.companyService.getCompanyByUserId(this.currentUserId).subscribe({
-        next: (res) => {
-          this.companyService.deleteCompany(res.id).subscribe({
-            next: () => {
-              this.authorizationService.logout()
+    let tiles = document.getElementsByClassName('tile');
+    for (let i= 0; i < tiles.length; i++) {
+      tiles.item(i)?.classList.add('d-none');
+    } 
+
+    let tileTitles = document.getElementsByClassName('tile-title')
+    for (let i= 0; i < tileTitles.length; i++) {
+      tileTitles.item(i)?.classList.add('d-none');
+    } 
+
+    document.getElementById('confirm-delete')?.classList.remove('d-none');
+   }
+
+   confirmCompanyDelete() {
+      if (confirm(this.translate.instant('Alert.deleteCompany'))) {
+        if (confirm(this.translate.instant('Alert.deleteCompanyConfirm'))) {
+          this.companyService.getCompanyByUserId(this.currentUser.id).subscribe({
+            next: (res) => {
+              this.companyService.deleteCompany(res.id).subscribe({
+                next: () => {
+                  this.authorizationService.logout()
+                }
+              })          
             }
-          })          
+          })
         }
-      })
     }
+   }
+
+   canModify() {
+    let canModify: boolean = false;
+
+    if (this.currentUser) {
+      this.canModifyRoles.forEach(role => {
+        if (this.currentUser.roles.includes(role)) {
+          canModify = true;
+        }
+      });
+    }
+
+    return canModify;
   }
+
 }
